New gadget: network-graph for network policy advisor
Inspektor Gadget's network-policy advisor relies on kprobes with
github.com/weaveworks/tcptracer-bpf/pkg/tracer. This is limited because
only TCP can be supported. Additionally, it sends a message to userspace
via the perf ring buffer and then to the kubectl-gadget plugin for every
connection, impacting performances.

This new implementation uses packet capture instead, making it possible
to support UDP. It uses BPF hash maps to reduce the overhead for each
connection.

This is implemented in the new gadget "network-graph"
(gadget=network-graph in the Trace custom resource) that gives a stream
of events. For now, the CLI still gather the log from each worker node.

The generation of network policies is improved in the following ways:
- Remove localhost policies: Kubernetes Network Policies don't block
  traffic to 127.0.0.1, so don't generate those policies
- Order policies in a consistent way for unit tests: this helps for unit
  tests and this is nicer for users.
- Fix testdata directory (the test data was in the wrong directory and
  therefore not tested)
- Use namespace label as defined by KEP-2161

Example of usage (local-gadget or Kubernetes): see
docs/guides/advise/network-policy.md

Integration tests:
  make kubectl-gadget
  cd integration/
  KUBECTL_GADGET=../kubectl-gadget go test ./... -integration -test.timeout 3m -run TestNetworkpolicy --no-deploy-spo -test.v=true

Unit tests:
  go test ./pkg/gadgets/networkpolicy/...