pkg/container-collection: replace container iteration with lookup maps
When enriching container information using the container-collection, we usually look up containers by either mntnsid
or netnsid. Previously, we returned matching containers after iterating over the main container map. With this patch,
we instead use dedicated lookup maps for both mntns and netns. Additionally, previously when looking up containers
using netnsid, a temporary array got created - this allocation has also been moved to the map maintenance and only
is done when adding or removing containers instead of on each lookup operation. This patch uses sync.RWMutex in
combination with standard maps, as in benchmarks it proved to be the most performant solution for this hot path.