pkg/gadget-service: migrate api to service
Previously, the gRPC API for `pkg/gadget-service` was defined inside `pkg/gadgettracermanager`
(like other API that the gadgettracermanager uses). However, as the service API is growing and we're
not only using it in gadgettracermanager anymore, it makes sense to move it to the gadget-service
and give it a dedicated .proto file.