pkg/gadget-service: add RunConfig and support for running over tcp sockets
This adds a RunConfig, which allows to configure the service to listen on
a tcp socket instead of just a previously used unix socket. In addition,
this now allows specifying a group id, which will then be used as the gid
for the unix socket to allow access to it using a usergroup.

Signed-off-by: Michael Friese <mfriese@microsoft.com>